background-remover/requirements.txt
numpy<=2.1.0
streamlit==1.41.1
rembg==2.0.62
Pillow==10.4.0
numba==0.61.0